# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"page",
					"limit"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"ErrorResponse": {
				"properties": {
					"error": {
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"message",
							"code"
						],
						"type": "object"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"ServiceType": {
				"type": "string",
				"enum": [
					"domestic",
					"commercial",
					"hotel",
					"laundry",
					"end_of_tenancy",
					"appliance"
				]
			},
			"PropertyCondition": {
				"type": "string",
				"enum": [
					"mildly_used",
					"heavily_used",
					"new"
				]
			},
			"PropertyType": {
				"type": "string",
				"enum": [
					"home",
					"condo",
					"office",
					"hotel"
				]
			},
			"BookingAddOn": {
				"type": "string",
				"enum": [
					"oven",
					"dryer",
					"airfryer",
					"dishwasher"
				]
			},
			"CreateBookingRequest": {
				"properties": {
					"address_id": {
						"type": "string"
					},
					"service_type": {
						"$ref": "#/components/schemas/ServiceType"
					},
					"condition": {
						"$ref": "#/components/schemas/PropertyCondition"
					},
					"property_type": {
						"$ref": "#/components/schemas/PropertyType"
					},
					"total_sq_metres": {
						"type": "number",
						"format": "double"
					},
					"rooms": {
						"type": "number",
						"format": "double",
						"description": "1-50"
					},
					"floors": {
						"type": "number",
						"format": "double",
						"description": "1-20"
					},
					"bathrooms": {
						"type": "number",
						"format": "double",
						"description": "0-20"
					},
					"add_ons": {
						"items": {
							"$ref": "#/components/schemas/BookingAddOn"
						},
						"type": "array"
					},
					"date": {
						"type": "string",
						"description": "ISO date, must be in the future"
					},
					"time": {
						"type": "string",
						"description": "HH:MM format"
					},
					"additional_info": {
						"type": "string",
						"description": "0-2000 chars"
					},
					"use_same_cleaner": {
						"type": "boolean"
					}
				},
				"required": [
					"address_id",
					"service_type",
					"condition",
					"property_type",
					"total_sq_metres",
					"rooms",
					"floors",
					"bathrooms",
					"date",
					"time"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BookingListItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"service_type": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"cleaner": {
						"properties": {
							"rating": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"rating",
							"name"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"service_type",
					"date",
					"time",
					"status",
					"cleaner"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListResponse_BookingListItem_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/BookingListItem"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"IdStatusResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"AdminListResponse_Record_string.unknown__": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Record_string.unknown_"
						},
						"type": "array"
					},
					"stats": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"stats",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"RoleListItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"display_name": {
						"type": "string"
					},
					"access": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"user_count": {
						"type": "number",
						"format": "double"
					},
					"date_added": {
						"type": "string"
					},
					"is_system": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"display_name",
					"access",
					"user_count",
					"date_added",
					"is_system"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RoleListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/RoleListItem"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateRoleResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"access": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"access"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateRoleRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Unique, 1-100 chars"
					},
					"display_name": {
						"type": "string"
					},
					"access": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"display_name",
					"access"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpdateRoleRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"display_name": {
						"type": "string"
					},
					"access": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ArchiveRoleResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"ReviewSnippet": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double"
					},
					"review_text": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"rating",
					"review_text",
					"created_at"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AdminCustomerDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"joined": {
						"type": "string"
					},
					"bookings_count": {
						"type": "number",
						"format": "double"
					},
					"last_booked": {
						"type": "string",
						"nullable": true
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/ReviewSnippet"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"rating",
					"joined",
					"bookings_count",
					"last_booked",
					"reviews"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AdminCreateBookingRequest": {
				"properties": {
					"customer_name": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"service_type": {
						"type": "string"
					},
					"condition": {
						"type": "string"
					},
					"property_type": {
						"type": "string"
					},
					"total_sq_metres": {
						"type": "number",
						"format": "double"
					},
					"rooms": {
						"type": "number",
						"format": "double"
					},
					"floors": {
						"type": "number",
						"format": "double"
					},
					"bathrooms": {
						"type": "number",
						"format": "double"
					},
					"add_ons": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"date": {
						"type": "string"
					},
					"time_start": {
						"type": "string"
					},
					"time_end": {
						"type": "string"
					},
					"cleaning_address": {
						"type": "string"
					},
					"additional_info": {
						"type": "string"
					},
					"use_same_cleaner": {
						"type": "boolean"
					}
				},
				"required": [
					"customer_name",
					"location",
					"email",
					"phone",
					"service_type",
					"rooms",
					"bathrooms",
					"date",
					"time_start",
					"cleaning_address"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"AdminCleanerDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"joined": {
						"type": "string"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/ReviewSnippet"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"rating",
					"joined",
					"reviews"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PayCleanerResponse": {
				"properties": {
					"transaction_id": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"transaction_id",
					"amount",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PayCleanerRequest": {
				"properties": {
					"amount": {
						"type": "number",
						"format": "double"
					},
					"account_number": {
						"type": "string"
					},
					"bank_name": {
						"type": "string"
					}
				},
				"required": [
					"amount",
					"account_number",
					"bank_name"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "fada-lustre-api",
		"version": "1.0.0",
		"description": "Backend API for Fada Lustre — a professional cleaning marketplace based in Derby, UK.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Fada Lustre",
			"email": "marketing@fadalustre.com"
		}
	},
	"paths": {
...
		"/v1/bookings": {
			"post": {
				"operationId": "CreateBooking",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"price": {
											"type": "string",
											"nullable": true
										},
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"price",
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new cleaning booking. Requires an address, service type, property details,\nand preferred date/time. The booking starts in 'unassigned' status.",
				"summary": "Create booking",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookingRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListBookings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_BookingListItem_"
								}
							}
						}
					}
				},
				"description": "List the authenticated customer's bookings with optional status filter\nand pagination. Returns bookings sorted by scheduled date.",
				"summary": "List my bookings",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/v1/admin/roles": {
			"get": {
				"operationId": "ListRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleListResponse"
								}
							}
						}
					}
				},
				"description": "List all admin roles with their assigned permissions.",
				"summary": "List roles",
				"tags": [
					"Admin Roles"
				],
				"security": [
					{
						"jwt": [
							"admin:control_permissions"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateRoleResponse"
								}
							}
						}
					},
					"409": {
						"description": "Duplicate",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Create a new admin role with a display name and set of permissions.",
				"summary": "Create role",
				"tags": [
					"Admin Roles"
				],
				"security": [
					{
						"jwt": [
							"admin:control_permissions"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoleRequest"
							}
						}
					}
				}
			}
		},
		"/v1/admin/roles/{id}": {
			"patch": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleListItem"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update a role's display name or permissions.",
				"summary": "Update role",
				"tags": [
					"Admin Roles"
				],
				"security": [
					{
						"jwt": [
							"admin:control_permissions"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoleRequest"
							}
						}
					}
				}
			}
		},
		"/v1/admin/roles/{id}/archive": {
			"patch": {
				"operationId": "ArchiveRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArchiveRoleResponse"
								}
							}
						}
					},
					"400": {
						"description": "Role has users",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Archive a role. Fails if any admin users are still assigned to it.",
				"summary": "Archive role",
				"tags": [
					"Admin Roles"
				],
				"security": [
					{
						"jwt": [
							"admin:control_permissions"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/admin/customers": {
			"get": {
				"operationId": "ListCustomers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminListResponse_Record_string.unknown__"
								}
							}
						}
					}
				},
				"description": "List all customers with filters for status, location, service type,\nand text search. Includes aggregate stats.",
				"summary": "List customers",
				"tags": [
					"Admin Customers"
				],
				"security": [
					{
						"jwt": [
							"admin:customers"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "service",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/admin/customers/{id}": {
			"get": {
				"operationId": "GetCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminCustomerDetail"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve detailed customer information including booking history,\nreviews, services booked, booking frequency, and last payment method.",
				"summary": "Get customer details",
				"tags": [
					"Admin Customers"
				],
				"security": [
					{
						"jwt": [
							"admin:customers"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/admin/customers/{id}/block": {
			"post": {
				"operationId": "BlockCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IdStatusResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Block a customer account, preventing them from logging in or making bookings.",
				"summary": "Block customer",
				"tags": [
					"Admin Customers"
				],
				"security": [
					{
						"jwt": [
							"admin:customers"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/admin/customers/{id}/book": {
			"post": {
				"operationId": "BookForCustomer",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"price": {
											"type": "number",
											"format": "double"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"status",
										"price",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a booking on behalf of a specific customer.",
				"summary": "Book for customer",
				"tags": [
					"Admin Customers"
				],
				"security": [
					{
						"jwt": [
							"admin:customers"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminCreateBookingRequest"
							}
						}
					}
				}
			}
		},
...
		"/v1/admin/cleaners": {
			"get": {
				"operationId": "ListCleaners",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminListResponse_Record_string.unknown__"
								}
							}
						}
					}
				},
				"description": "List all cleaners with filters for status, location, service type,\nand text search. Includes aggregate stats.",
				"summary": "List cleaners",
				"tags": [
					"Admin Cleaners"
				],
				"security": [
					{
						"jwt": [
							"admin:cleaners"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "service",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/admin/cleaners/{id}": {
			"get": {
				"operationId": "GetCleaner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminCleanerDetail"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve detailed cleaner information including reviews, availability,\npayment history, and performance metrics.",
				"summary": "Get cleaner details",
				"tags": [
					"Admin Cleaners"
				],
				"security": [
					{
						"jwt": [
							"admin:cleaners"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/admin/cleaners/{id}/block": {
			"post": {
				"operationId": "BlockCleaner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IdStatusResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Block a cleaner account, preventing them from receiving bookings.",
				"summary": "Block cleaner",
				"tags": [
					"Admin Cleaners"
				],
				"security": [
					{
						"jwt": [
							"admin:cleaners"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/admin/cleaners/{id}/pay": {
			"post": {
				"operationId": "PayCleaner",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PayCleanerResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Process a manual payment to a cleaner's bank account.\nCreates a transaction record for audit tracking.",
				"summary": "Pay cleaner",
				"tags": [
					"Admin Cleaners"
				],
				"security": [
					{
						"jwt": [
							"admin:cleaners"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PayCleanerRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}